refactor: extract scale rules
